% Generated by roxygen2 (4.0.2): do not edit by hand
\name{changeColName}
\alias{changeColName}
\title{changeColName()}
\usage{
changeColName(data, current_colname, desired_colname)
}
\arguments{
\item{data}{a data.frame or data.table}

\item{current_colname}{the column name that you want to change}

\item{desired_colname}{the column name that you current_colname changed to}
}
\description{
Changes a column name from one to another. Accepts both data frames and
data.tables.
}
\details{
Note that this function has side effects, namely, it modifies
an object in its calling environment. This was necessary for data.tables, as
the goal is to modify them in place. I decided to do the same for data
frames to make the function syntax the same, even though R won't do
modification in place for data frames.
}
\examples{
Here is how to change a column name, where we change a column named "Jake"
into one named "Josh":

changeColName(my_data, "Jake", "Josh")
}
\keyword{crosswalk}
\keyword{lookup,}
\keyword{match,}
\keyword{matchup,}
\keyword{up,}

