% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/miscFunctions.R
\name{changeObjectName}
\alias{changeObjectName}
\title{changeObjectName()}
\usage{
changeObjectName(oldObj, newObj, overwriteExisting = TRUE)
}
\arguments{
\item{oldObj}{the object that should have its name changed}

\item{newObj}{the new name of the object}

\item{overwriteExisting}{TRUE by default, should the function overwrite
existing objects with the name newObj?}
}
\description{
Change the name of an object. Uses non-standard evaluation (NSE) to capture
the unevaluted expressions given for its arguments (this should become clear
in the examples below.) See changeObjectName_ for a version of this function
that doesn't use NSE.
}
\examples{
jake <- c(5, 6, 7)
changeObjectName(jake, john)
}

