% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/miscFunctions.R
\name{changeObjectName_}
\alias{changeObjectName_}
\title{changeObjectName_()}
\usage{
changeObjectName_(oldObj, newObj, overwriteExisting = TRUE)
}
\arguments{
\item{oldObj}{the object that should have its name changed}

\item{newObj}{the new name of the object}
}
\description{
Change the name of an object. Uses standard evaluation (SE) to capture
to change the name of an object given character vectors for the current
name of an object and the desired name.
}
\examples{
jake <- c(5, 6, 7)
changeObjectName("jake", "john")
}

