% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createDelimited}
\alias{createDelimited}
\title{createDelimited()}
\usage{
createDelimited(table, delimit_this, by_this, delimiter = ", ",
  filter = NULL, returnDT = NULL)
}
\arguments{
\item{table}{a data.frame or data.table}

\item{delimit_this}{the column that you want to become a delimited field}

\item{by_this}{the Id column which seperates different delimited sections}

\item{delimiter}{which delimiter do you want to use? Default is ", "}

\item{filter}{a data.table i expression (roughly equivalent to dplyr's
filter function) to filter the data before creating delimited fields}

\item{returnDT}{NULL by default, can set to TRUE or FALSE to explicitly
return a data.table. By default, a data.table is returned if table is a
data.table, and a data.frame is returned if table is a data.frame.}
}
\description{
Turns long data into delimited data. Ex., imagine you have a table with 2
columns, the first is an Id column with "A" repeated 3 times, the second
column has rows X then Y then Z. This function would turn this table from
having three rows into having one row where X, Y, Z are delimited by
some delimiter (the default is ", " like the above example).
}
\keyword{crosswalk}
\keyword{lookup,}
\keyword{match,}
\keyword{matchup,}
\keyword{up,}

