% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createDirIfNotExist.R
\name{createDirIfNotExist}
\alias{createDirIfNotExist}
\title{createDirIfNotExist()}
\usage{
createDirIfNotExist(dir_location, ...)
}
\arguments{
\item{dir_location}{the directory you want to create}

\item{...}{additional arguments to be passed to dir.create}
}
\description{
Creates a directory if it doesn't already exist.
}
\keyword{create,}
\keyword{dir,}
\keyword{directory,}
\keyword{exist,}
\keyword{existing}
\keyword{if,}
\keyword{not,}

