% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/createMatchup.R
\name{createMatchup}
\alias{createMatchup}
\title{createMatchup()}
\usage{
createMatchup(matching_data, match_this, to_this)
}
\arguments{
\item{matching_data}{a data frame containing columns matching two things,
i.e. the column names you are inputting in the match_this and to_this
arguments.}

\item{match_this}{column name of what you want matched (ex. state initials)...}

\item{to_this}{...column name of what you want the above matched to (ex. state names)}
}
\description{
Creates a closure of the matchup() function that stores with the matchup()
function data for all of its arguments except for its first, the vector of
elements that need to be matched. This allows one to create quick, 1:1
matching functions that can be reused.
}
\examples{
Let's try using createMatchup to create a function that matches US state
initials to state names, like the example for the matchup() function.
Remember, this is how we matched state initials to names before:

matchup(vector_of_state_initials, us_states, "state", "state_name")

Now let's create a function to do this, called initials_to_names():

initals_to_names <- createMatchup(us_states, "state", "state_name")

Now, we just need to input the vector_of_state_initials that was the first
argument in the matchup() function above to produce its output when it comes
to matching state names to initials. This:

initials_to_names(vector_of_state_initials)

Produces the same output as this:

matchup(vector_of_state_initials, us_states,
                       "state", "state_name")
}
\keyword{crosswalk}
\keyword{lookup,}
\keyword{match,}
\keyword{matchup,}
\keyword{up,}

