% Generated by roxygen2 (4.0.2): do not edit by hand
\name{factorToInteger}
\alias{factorToInteger}
\title{factorToInteger()}
\usage{
factorToInteger(factors, intCheck = TRUE)
}
\arguments{
\item{factors}{a vector of factors}

\item{intCheck}{samples elements from factors to determine whether or not
they are whole numbers. If they are not whole numbers, factorToNumeric is
used instead and a warning shows up. Default is \code{TRUE}, set this
argument to \code{FALSE} to disable this behavior.}
}
\description{
Turns factors into numerics.
}
\examples{
factors <- factor(c(1,2,3,4,5))
factors <- factorToInteger(factors)
}
\keyword{factor,}
\keyword{numeric}

