% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datatableSpecific.R
\name{keepColsDT}
\alias{keepColsDT}
\title{keepColsDT()}
\usage{
keepColsDT(data = NULL, keepCols = NULL, verbose = TRUE)
}
\arguments{
\item{data}{a data.table}

\item{keepCols}{a character vector of column names NOT to delete}

\item{verbose}{TRUE by default, if FALSE it turns off the messages that
appear after each column is deleted by reference.}
}
\description{
Deletes all of the columns in a data.table by reference except for the
columns that you specify to keep.
}
\examples{
columnsToKeep <- c("mpg", "cyl", "disp")
mtcars <- data.table(mtcars)
keepColsDT(mtcars, columnsToKeep)
}
\keyword{DT,}
\keyword{columns,}
\keyword{data.table}
\keyword{keep,}

