% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datatableSpecific.R
\name{mergeDT}
\alias{mergeDT}
\title{mergeDT()}
\usage{
mergeDT(DT1 = NULL, DT2 = NULL, keys = NULL, keepCols = NULL,
  deleteCols = NULL)
}
\arguments{
\item{DT1}{a data.table}

\item{DT2}{a data.table}

\item{keys}{a character vector of one or more columns to join on between
DT1 or DT2. If you have already used \code{data.table::setkey()} to set
the keys for DT1 and DT2 and those keys are identical, and you do not
specify the keys argument, the join will be performed on those columns.}

\item{keepCols}{a character vector of column names NOT to delete. You may
specify keepCols or deleteCols but NOT BOTH.}

\item{deleteCols}{a character vector of column names to delete. You may
specify keepCols or deleteCols but NOT BOTH.}
}
\description{
Merges two data.tables using
}
\examples{
Merging two data.tables where we specify the key in \\code{mergeDT}:
mergeDT(DT1, DT2, keys = "common_id")

Setting keys first, then merging two data.tables:
setkey(DT1, common_id)
setkey(DT2, common_id)
mergeDT(DT1, DT2)
}
\keyword{DT,}
\keyword{data.table,}
\keyword{fast,}
\keyword{key,}
\keyword{merge,}
\keyword{setkey}
\keyword{tables,}

