% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NAs.R
\name{NaNs}
\alias{NaNs}
\title{NaNs()}
\usage{
NaNs(data, type, digits)
}
\arguments{
\item{data}{an object, can be a data.frame (see NaNs by column), matrix (also
see NaNs by column), or a vector}

\item{type}{\code{"number"}  is the default, displaying the number of NaNs.
You may also use \code{"percent"}  to see the percent of NaNs}

\item{digits}{the number of digits to round to when the type arguemnt is set
to \code{"percent"}}
}
\description{
Returns the number of NaNs in an object by column (for data.frames and
matricies) or just the number of NaNs (for all other objects). You may also
set the type argument to \code{"percent"} to see the percent of NAs.
}
\examples{
To see the number of NaNs either in total or by column, depending on the
class of the object inputted:
\\code{NaNs(my_data)}

Same as above, but now looking at the percent of NAs:
\\code{NaNs(my_data, type = "percent")}
}
\keyword{NaN,}
\keyword{NaNs,}
\keyword{data}
\keyword{missing,}
\keyword{percent,}
\keyword{valid,}

