% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NAs.R
\name{NAs}
\alias{NAs}
\title{NAs()}
\usage{
NAs(data, type, digits)
}
\arguments{
\item{data}{an object, can be a data.frame (see NAs by column), matrix (also
see NAs by column), or a vector}

\item{type}{\code{"number"}  is the default, displaying the number of NAs.
You may also use \code{"percent"}  to see the percent of NAs}

\item{digits}{the number of digits to round to when the type arguemnt is set
to \code{"percent"}}
}
\description{
Returns the number of NAs in an object by column (for data.frames and
matricies) or just the number of NAs (for all other objects). You may also
set the type argument to \code{"percent"} to see the percent of NAs.
}
\examples{
To see the number of NAs either in total or by column, depending on the class
of the object inputted:
\\code{NAs(my_data)}

Same as above, but now looking at the percent of NAs:
\\code{NAs(my_data, type = "percent")}
}
\keyword{NA,}
\keyword{NAs,}
\keyword{data,}
\keyword{gone}
\keyword{missing,}
\keyword{percent,}
\keyword{valid,}

