% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NAs.R
\name{NonNAs}
\alias{NonNAs}
\title{NonNAs()}
\usage{
NonNAs(data, type, digits)
}
\arguments{
\item{data}{an object, can be a data.frame (see Non-NAs by column), matrix
(also see Non-NAs by column), or a vector}

\item{type}{\code{"number"}  is the default, displaying the number of
Non-NAs. You may also use \code{"percent"}  to see the percent of Non-NAs}

\item{digits}{the number of digits to round to when the type arguemnt is set
to \code{"percent"}}
}
\description{
Returns the number of Non-NAs in an object by column (for data.frames and
matricies) or just the number of Non-NAs (for all other objects). You may also
set the type argument to \code{"percent"} to see the percent of Non-NAs.
}
\examples{
To see the number of Non-NAs either in total or by column, depending on the
class of the object inputted:
\\code{NonNAs(my_data)}

Same as above, but now looking at the percent of Non_NAs:
\\code{NonNAs(my_data, type = "percent")}
}
\keyword{NA,}
\keyword{NAs,}
\keyword{data,}
\keyword{gone}
\keyword{missing,}
\keyword{percent,}
\keyword{valid,}

