% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/miscFunctions.R
\name{splitRmNA}
\alias{splitRmNA}
\title{splitRmNA()}
\usage{
splitRmNA(vector1, vector2, ...)
}
\arguments{
\item{vector1}{}

\item{vector2}{}

\item{...}{additional arguments to be passed to \code{base::split}}
}
\description{
Splits one vector by another vector after removing elements from both
vectors when either vector is NA.
}
\keyword{NA,}
\keyword{NAs}
\keyword{remove,}
\keyword{split,}
\keyword{vector,}

