% Generated by roxygen2 (4.0.2): do not edit by hand
\name{valuesToValue}
\alias{valuesToValue}
\title{valuesToValue()}
\usage{
valuesToValue(data = NULL, values = NULL, valueToConvertTo = NULL,
  onlyConvert = NULL, noConvert = NULL)
}
\arguments{
\item{data}{a data frame/table that we want to remove NAs from}

\item{values}{a vector of one or more values that you wish to be converted
into another value}

\item{valueToConvertTo}{the value that you want values converted to}

\item{onlyConvert}{optional - specify a vector of one or more column names
as the only columns where the NA conversion will take place. You may either
have this parameter satified, or the noConvert parameter specified, you may
not specify both.}

\item{noConvert}{optional - specify a vector of one or more columns names
that you do not want any NA conversion applied to. You may either have this
parameter satified, or the noConvert parameter specified, you may not
specify both.}
}
\description{
Given a data.frame or data.table, turns selected values into a different value for
specified columns. By default all columns are used, but you may specify the
onlyConvert argument with a vector of one or more columns names and only do
the conversion on these columns, or alternatively, specify the noConvert
argument with a vector of one or more columns names that shouldnt have the
conversion done on them. You may only choose to specify onlyConvert or
noConvert, you may not specify both at the same time.
}
\details{
Note that this function has side effects, namely, it modifies
an object in its calling environment. This was necessary for data.tables, as
the goal is to modify them in place. I decided to do the same for data
frames to make the function syntax the same, even though R wont do
modification in place for data frames.
}
\examples{
Lets say we want to convert "na" or -500 values into "jake" for only the
columns "town" and "city" in my_data:
valuesToNA(my_data, c("na", -500), "jake" ,onlyConvert = c("town", "city"))

Or, what if we want to convert "na" or -500 values into "jake" for every
columnn in my_data except for "town", "city", or "country":
valuesToNA(my_data, c("na", -500), "jake", noConvert = c("town", "city",
"country"))
}
\keyword{valuesToValue}

