% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/changeColName.R
\name{changeColName}
\alias{changeColName}
\title{changeColName()}
\usage{
changeColName(data, ..., ref = TRUE, warnings = TRUE, invisible = TRUE)
}
\arguments{
\item{data}{a \code{data.frame} or \code{data.table}}

\item{...}{either a) if you are only changing one column, the current column
name, followed by a comma and then the desired column name (option of NSE),
or b) one or more of the following: current_column_name/desired_column_name,
where the current column name is followed by a forward slash followed by
your desired column name. Using the function in this way, you may change
as many column names as you like in one function call.}

\item{ref}{TRUE (default) or FALSE, if TRUE and data is a data.table, modify
the data.table by reference (modifying-in-place), if FALSE, do not modify
the data.table by reference, instead treat it like a data.frame (copy on
modify).}

\item{warnings}{TRUE (default) or FALSE, should warnings occur when
modifications by reference occur or conversions take place?}

\item{invisible}{TRUE (by default), invisibly return the data?}
}
\description{
Changes a column name from one to another. Accepts both data frames and
data.tables. Multiple column names can be changed with the simple syntax
\code{old_column_name/new_column_name}.
}
\details{
This function uses seperate methods for data.tables and data.frames. By
default, data.tables will be modified by reference. To turn off this
behavior, set \code{ref} to \code{FALSE}. Resulting \code{data.frames/tables}
will be invisibly returned.
}
\examples{
Here is how to change a column name, where we change a column named "Jake"
into one named "Josh":

\\code{changeColName(my_data, "Jake", "Josh")}
}
\keyword{crosswalk}
\keyword{lookup,}
\keyword{match,}
\keyword{matchup,}
\keyword{up,}

