% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datatableSpecific.R
\name{deleteColsDT}
\alias{deleteColsDT}
\title{deleteColsDT()}
\usage{
deleteColsDT(data = NULL, deleteCols = NULL, verbose = TRUE)
}
\arguments{
\item{data}{a data.table}

\item{deleteCols}{a character vector of column names o delete}

\item{verbose}{TRUE by default, if FALSE it turns off the messages that
appear after each column is deleted by reference.}
}
\description{
Deletes all of the specified columns in a data.table by reference.
}
\examples{
columnsToDelete <- c("mpg", "cyl", "disp")
mtcars <- data.table(mtcars)
deleteColsDT(mtcars, columnsToDelete)
}
\keyword{DT,}
\keyword{columns,}
\keyword{data.table}
\keyword{delete,}

