% Generated by roxygen2 (4.0.2): do not edit by hand
\name{factorToNumeric}
\alias{factorToNumeric}
\title{factorToNumeric()}
\usage{
factorToNumeric(factors)
}
\arguments{
\item{a}{vector of factors}
}
\description{
Turns factors into numerics.
}
\examples{
factors <- factorToNumeric(factors)
}
\keyword{factor,}
\keyword{numeric}

