% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/factorToNumeric.R
\name{factorToNumeric}
\alias{factorToNumeric}
\title{factorToNumeric()}
\usage{
factorToNumeric(factors)
}
\arguments{
\item{a}{vector of factors}
}
\description{
Turns factors into numerics.
}
\examples{
factors <- factorToNumeric(factors)
}
\keyword{factor,}
\keyword{numeric}

