% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/factorToNumeric.R
\name{factorToNumeric}
\alias{factorToNumeric}
\title{factorToNumeric()}
\usage{
factorToNumeric(factors)
}
\arguments{
\item{factors}{factors to convert to numeric}
}
\description{
Turns factors into numerics.
}
\examples{
factors <- factor(c(1.2, 4.2, 4.2, 3.4))
factors <- factorToNumeric(factors)
}
\keyword{factor,}
\keyword{numeric}

