% Generated by roxygen2 (4.0.2): do not edit by hand
\name{keepColsDT}
\alias{keepColsDT}
\title{keepColsDT()}
\usage{
keepColsDT(data = NULL, keepCols = NULL, silent = FALSE)
}
\arguments{
\item{data}{a data.table}

\item{keepCols}{a character vector of column names NOT to delete}

\item{slient}{FALSE by default, if TRUE it turns off the messages that
appear after each column is deleted by reference.}
}
\description{
Deletes all of the columns in a data.table by reference except for the
columns that you specify to keep.
}
\examples{
columnsToKeep <- c("mpg", "cyl", "disp")
mtcars <- data.table(mtcars)
keepColsDT(mtcars, columnsToKeep)
}
\keyword{DT,}
\keyword{columns,}
\keyword{data.table}
\keyword{keep,}

