% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/miscFunctions.R
\name{listToObjects}
\alias{listToObjects}
\title{listToObjects()}
\usage{
listToObjects(myList, deleteList = TRUE, objectNames = NULL,
  overwriteExisting = TRUE, verbose = TRUE)
}
\arguments{
\item{myList}{a list}

\item{deleteList}{TRUE by default, should the list be deleted after?}

\item{verbose}{TRUE by default, should the function be verbose?}
}
\description{
Turns each element of a list into an object in your global environment with
the name of each element in the list becoming the name of that object. By
default, the list is deleted at the end, although this behavior may be
turned off.
}
\examples{
jake <- list(a = c(1, 2, 3), b = c(2, 3, 4))
listToObjects(jake)
}
\keyword{into,}
\keyword{list,}
\keyword{objects}
\keyword{transform,}

