% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datatableSpecific.R
\name{subsetDT}
\alias{subsetDT}
\title{subsetDT()}
\usage{
subsetDT(data, ..., ref = TRUE)
}
\arguments{
\item{data}{a data.table}

\item{...}{one or more lists, where the first element of the list is the
column name, and the second element is one or more items from that column
to subset.}

\item{ref}{TRUE by default, it will take the keys that are set during
the subset process and apply those keys to data by reference. If FALSE,
a copy of data is made via \code{data.table::copy()}, leaving data
unaffected by the key setting in subsetDT.}
}
\description{
Subsets a data.table by either setting keys, or using keys that have already
been set. This type of subset may be faster than using a subset expression
in the i clause of a data.table statement without setting keys.
}
\keyword{DT,}
\keyword{data.table}
\keyword{rows,}
\keyword{subset,}

